+++ src/main/java/it/academy/dao/DeveloperDao.java
@@ -12,10 +12,10 @@ public interface DeveloperDao extends Dao<Developer, Long> {
 
     List<Developer> getDevelopers(UserStatus status, int page, int count) throws NoResultException, IOException;
 
-    Integer getCountOfDevelopers(UserStatus status) throws NoResultException, IOException;
+   Long getCountOfDevelopers(UserStatus status) throws NoResultException, IOException;
 
     List<Developer> getDevelopersByContractorId(Long contractorId, ProjectStatus status, int page, int count)
         throws NoResultException, IOException;
 
-    Integer getCountOfDevelopers(Long contractorId, ProjectStatus status) throws NoResultException, IOException;
+    Long getCountOfDevelopers(Long contractorId, ProjectStatus status) throws NoResultException, IOException;
 }
